requirements.txt
anyio==4.4.0
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.10.0
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
eyeD3==0.8.10
fastjsonschema==2.20.0
filelock==3.6.0
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
mutagen==1.45.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemo-emblems==6.0.1
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
oauthlib==3.2.0
onboard==1.4.1
overrides==7.7.0
packaging==24.1
PAM==0.4.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-magic==0.4.24
python-telegram-bot==21.4
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
qrcode==7.3.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
SecretStorage==3.3.1
Send2Trash==1.8.3
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
stack-data==0.6.3
systemd-python==234
terminado==0.18.1
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
uri-template==1.3.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==9.1
widgetsnbextension==4.0.11
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
